requirements.txt
python-telegram-bot==22.2
flask==3.1.1
httpx==0.28.1
nest_asyncio==1.6.0